doc cleanup
