+++ .gitignore
@@ -42,6 +42,7 @@ ehthumbs.db
 Thumbs.db
 .spyderworkspace
 *.*~
+*~
 
 # Python pyproducts #
 #####################
