+++ README.md
@@ -1,2 +1,4 @@
 # data-quality
-The Data Quality project will research, evaluate, and then design and implement solution to ensure qulity od scientific data.
+The Data Quality project will research, evaluate, and then design and implement solution to ensure quality of scientific data collected at the Advanced Photon Source.
+
+Project Home Page: http://dquality.readthedocs.org/
