+++ doc/index.rst
@@ -2,10 +2,12 @@
 
 .. image:: source/img/dquality-logo.png
    :width: 320px
-   :alt: DMagic
+   :alt: DQuality
+
+
 
 `DQuality <https://github.com/bfrosik/data-quality>`_ is an open-sourced Python
-toolbox to perform data quality assessment.
+toolbox to perform data quality assessment at the Advanced Photon Source.
 
 This guide is maintained on 
 `GitHub <https://github.com/bfrosik/data-quality/tree/master/doc>`_.
@@ -14,8 +16,10 @@ This guide is maintained on
    :maxdepth: 1
    
    source/about
+   source/deliverable
    source/install
    source/devguide
+   source/api
    source/faq
