+++ doc/source/api.rst
@@ -1,0 +1,16 @@
+API reference
+=============
+
+This section contains the API reference and usage information
+for `DQuality <https://github.com/bfrosik/data-quality>`_.
+
+.. rubric:: **Quality Modules:**
+
+.. toctree::
+
+   api/dquality.verifier
+
+.. automodule:: dquality
+   :members:
+   :undoc-members:
+   :show-inheritance:
