+++ doc/source/api/dquality.verifier.rst
@@ -1,0 +1,17 @@
+
+:mod:`dquality.verifier`
+========================
+
+.. automodule:: dquality.verifier
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+   .. rubric:: **Functions:**
+
+   .. autosummary::
+   
+      validateMeansignalIntensity
+      validateSignalIntensityStandardDeviation
+      verifyFileQuality
+      monitorDir
