+++ manage/project-outline
@@ -1,4 +1,4 @@
-The project will have the following major deliverables:
+The assesment of raw and processed data quality is essential to provide conistency, accuracy, reliability and reproducibility of scientific results. This project has the following major deliverables:
 
 1. Prototype (Data Quality Verifier, DQV)
 It will be a stand alone application.
